refactor(pixelmap): WritePixels/ReadPixels YUV 分支 readSize 改用 uint64_t
ReadPixels 与 WritePixels 的 IsYUV 分支中 readSize 原为有符号 int，与周围
tmpSize(uint64_t)、pixelsSize_(uint32_t)、bufferSize(uint64_t)、
MAX_READ_COUNT(uint32_t) 的无符号算术混用：memcpy_s 的 count 参数存在
int→size_t 有符号转无符号隐患（仅靠 readSize 恒为正这一不变量兜底），
(int)(pixelsSize_ - tmpSize) / (int)(bufferSize - tmpSize) 为 uint64→int
收窄转换（仅靠分支把值限到 <2048 才安全）。改为 uint64_t readSize，直接
赋值 pixelsSize_ - tmpSize / bufferSize - tmpSize（同为 uint64），去掉
(int) 收窄与 static_cast<uint64_t>，由类型堵死负值隐患。

零行为变化：readSize 取值恒在 [1,2048]，改前改后数值逐位相同，memcpy_s
的 count 与 tmpSize 累加值不变。



AI[0%] Human Fixed[0%] Human[100%] AI Adopted[0%]

+++ frameworks/innerkitsimpl/common/src/pixel_map.cpp
@@ -2038,19 +2038,19 @@ uint32_t PixelMap::ReadPixels(const uint64_t &bufferSize, uint8_t *dst)
     }
     if (IsYUV(imageInfo_.pixelFormat)) {
         uint64_t tmpSize = 0;
-        int readSize = MAX_READ_COUNT;
+        uint64_t readSize = MAX_READ_COUNT;
         while (tmpSize < bufferSize && tmpSize < pixelsSize_) {
             if (tmpSize + MAX_READ_COUNT > bufferSize) {
-                readSize = (int)(bufferSize - tmpSize);
+                readSize = bufferSize - tmpSize;
             } else if (tmpSize + MAX_READ_COUNT > pixelsSize_) {
-                readSize = (int)(pixelsSize_ - tmpSize);
+                readSize = pixelsSize_ - tmpSize;
             }
             errno_t ret = memcpy_s(dst + tmpSize, readSize, data_ + tmpSize, readSize);
             if (ret != 0) {
                 IMAGE_LOGE("read pixels by buffer memcpy the pixelmap data to dst fail, error:%{public}d", ret);
                 return ERR_IMAGE_READ_PIXELMAP_FAILED;
             }
-            tmpSize += static_cast<uint64_t>(readSize);
+            tmpSize += readSize;
         }
     } else {
         // Copy the actual pixel data without padding bytes
@@ -2433,17 +2433,17 @@ uint32_t PixelMap::WritePixels(const uint8_t *source, const uint64_t &bufferSize
 
     if (IsYUV(imageInfo_.pixelFormat)) {
         uint64_t tmpSize = 0;
-        int readSize = MAX_READ_COUNT;
+        uint64_t readSize = MAX_READ_COUNT;
         while (tmpSize < bufferSize && tmpSize < pixelsSize_) {
             if (tmpSize + MAX_READ_COUNT > pixelsSize_) {
-                readSize = (int)(pixelsSize_ - tmpSize);
+                readSize = pixelsSize_ - tmpSize;
             }
             errno_t ret = memcpy_s(data_ + tmpSize, pixelsSize_ - tmpSize, source + tmpSize, readSize);
             if (ret != 0) {
                 IMAGE_LOGE("write pixels by buffer memcpy the pixelmap data to dst fail, error:%{public}d", ret);
                 return ERR_IMAGE_WRITE_PIXELMAP_FAILED;
             }
-            tmpSize += static_cast<uint64_t>(readSize);
+            tmpSize += readSize;
         }
     } else {
         for (int i = 0; i < imageInfo_.size.height; ++i) {
